Update SpritesheetInstancing.cs
Changed Array.Resize to creating a new array in the Begin() method. Resizing is no longer necessary because the array is completely rebuilt at the start of each Begin() call.
Now, the Garbage Collector will only need to work if the numberOfElements changes. If the number of elements remains the same in subsequent Begin() calls for this instance, the array will not be garbage collected, improving performance.