Update README.md
New Info about the rezising of the array with multiple Begin() Methods and different amounts of elements per DrawCall:

"If a single instance of the Spritesheet Instancing class is used across multiple Begin() calls, it is recommended to set the numberOfElements to the same value for each call. Ideally, use a size equal to or greater than the maximum expected number of elements. This prevents the class from creating new arrays every frame, reducing unnecessary allocations and minimizing work for the Garbage Collector."